Removing the code to read from the serial port. Use stdin instead.
This change allows to make this code independent of the way the data
was gathered from the multimeter.

This allows to adapt it to newer versions of the multimeter which
don't come with a cable providing a virtual serial port but instead
come with a cable providing a USB HID device.

With this change, the data analysis is now independent of the data taking:
You simply pipe the output of the multimeter into this utility
and it will interpret it for you.

This makes it more like the es51922 tool from ut61e-linux-sw-0.02.tar.gz
found on http://diyftw.de/wiki/doku.php?id=projekte:ut61e .